+++ detectors/node/opentelemetry-resource-detector-aws/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.30.0 to ^0.31.0
+
 ## [1.1.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-aws-v1.0.3...resource-detector-aws-v1.1.0) (2022-05-14)