+++ detectors/node/opentelemetry-resource-detector-aws/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.30.0 to ^0.31.0
+
+## [1.1.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-aws-v1.0.3...resource-detector-aws-v1.1.0) (2022-05-14)
+
+
+### Features
+
+* add supported node versions for all packages ([#973](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/973)) ([baaacbd](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/baaacbdd35ca4baab0afae64647aa8c0380ee4b7))
+* use Otel SDK 1.2/0.28 ([#984](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/984)) ([098c2ed](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/098c2ed6f9c5ab7bd865685018c0777245aab3b7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from 0.29.0 to ^0.30.0
+
 ### [1.0.3](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-aws-v1.0.2...resource-detector-aws-v1.0.3) (2022-01-24)